Create BaseQuery and add an example
