+++ app/graphql/queries/base_query.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Queries
+  class BaseQuery < GraphQL::Schema::Resolver
+  end
+end
