+++ app/graphql/queries/survey_query.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Queries
+  class SurveyQuery < BaseQuery
+    type [Types::SurveyType], null: false
+    description 'Survey list'
+
+    def resolve
+      Survey.all
+    end
+  end
+end
