--- app/graphql/types/user_type.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  class UserType < Types::Base::Object
-    field :id, ID, null: false
-    field :email, String, null: false
-  end
-end
